Examination: Permissing edit record (Symptom, Sign, Image, Exploration)

+++ app/Http/Controllers/Staff/RecordSignController.php
@@ -7,28 +7,43 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Record_Sign;
+use App\Record;
+use Auth;
 use DB;
 class RecordSignController extends Controller
 {
     public function update($record_id, Request $request){
-		$addArrRaw = json_decode($request->addArr, true);
-		$addArr = [];
-		$deleteArr = json_decode($request->deleteArr, true);
-		if(count($addArrRaw) > 0)
+    	$record = Record::find($record_id);
+		if($record->state==0||($record->state==1&&$record->examiner==Auth::user()->id))
 		{
-			foreach($addArrRaw as $item)
-				array_push($addArr, ['record_id' => $record_id, 'sign_id' => $item]);
-			Record_Sign::insert($addArr);
-		}
-		if(count($deleteArr) > 0)
-			Record_Sign::destroy($deleteArr);
+			$addArrRaw = json_decode($request->addArr, true);
+			$addArr = [];
+			$deleteArr = json_decode($request->deleteArr, true);
+			if(count($addArrRaw) > 0)
+			{
+				foreach($addArrRaw as $item)
+					array_push($addArr, ['record_id' => $record_id, 'sign_id' => $item]);
+				Record_Sign::insert($addArr);
+			}
+			if(count($deleteArr) > 0)
+				Record_Sign::destroy($deleteArr);
 
-		//Response
-		$signs = DB::table('signs')
-		            ->leftJoin(DB::raw("(SELECT * FROM record_sign WHERE record_id = $record_id) AS temp_tbl"), 'signs.id', '=', 'temp_tbl.sign_id')
-		            ->select('signs.id AS index_id', 'signs.name', 'temp_tbl.id AS id')
-		            ->orderBy('index_id')
-		            ->get();
-        return ['state' => 1, 'list' => $signs];
+			//Response
+			$signs = DB::table('signs')
+			            ->leftJoin(DB::raw("(SELECT * FROM record_sign WHERE record_id = $record_id) AS temp_tbl"), 'signs.id', '=', 'temp_tbl.sign_id')
+			            ->select('signs.id AS index_id', 'signs.name', 'temp_tbl.id AS id')
+			            ->orderBy('index_id')
+			            ->get();
+	        return ['state' => 1, 'list' => $signs, 'message' => 'Đã cập nhật thành công!'];
+	    }
+	    else
+		{
+			$signs = DB::table('signs')
+			            ->leftJoin(DB::raw("(SELECT * FROM record_sign WHERE record_id = $record_id) AS temp_tbl"), 'signs.id', '=', 'temp_tbl.sign_id')
+			            ->select('signs.id AS index_id', 'signs.name', 'temp_tbl.id AS id')
+			            ->orderBy('index_id')
+			            ->get();
+			return ['state' => 0, 'list' => $signs, 'message' => 'Bạn không đủ quyền để cập nhật bệnh án này'];
+		}
 	}
 }
